feat(cmake): centralize compiler warnings with ROS2MedkitWarnings module
Add ROS2MedkitWarnings.cmake with production-grade warning flags
aligned with OpenSSF, AUTOSAR C++14, and MISRA C++ guidelines.
Enable -Werror by default. Replace per-package add_compile_options
blocks with a single include(ROS2MedkitWarnings) across all 12
packages.

New flags: -Wswitch-enum, -Wfloat-equal, -Wcast-qual, -Wundef,
-Wzero-as-null-pointer-constant, -Wextra-semi, -Wdouble-promotion,
-Wsign-conversion, -Wnon-virtual-dtor, -Wold-style-cast,
-Woverloaded-virtual, -Wcast-align, -Wnull-dereference,
-Wimplicit-fallthrough, -Wformat=2, -Wuseless-cast (GCC),
-Wduplicated-cond, -Wduplicated-branches, -Wlogical-op (GCC).

Fix all warnings surfaced by the new flags: implicit conversions,
useless casts, incomplete switch-enum coverage, float-to-double in
printf, char sign conversions.

Exclude vendored dynmsg and gtest/gmock from strict warnings via
set_source_files_properties and ros2_medkit_relax_vendor_warnings().

Closes #356

+++ src/ros2_medkit_gateway/src/http/handlers/sse_fault_handler.cpp
@@ -52,7 +52,7 @@ SSEFaultHandler::SSEFaultHandler(HandlerContext & ctx, std::shared_ptr<SSEClient
   if (keepalive_interval <= std::chrono::milliseconds::zero()) {
     RCLCPP_WARN(HandlerContext::logger(),
                 "Non-positive SSE keepalive override %" PRId64 "ms rejected; using default %ds",
-                static_cast<int64_t>(keepalive_interval.count()), kKeepaliveIntervalSec);
+                keepalive_interval.count(), kKeepaliveIntervalSec);
   }
 
   const auto fault_events_topic = build_fault_manager_events_topic(ctx_.node());
